pipeline/producer/requirements.txt
Faker
kafka-python
python-dotenv